Add new Folder class for folder abstraction
